refactor settings into sidebar as a override context
