+++ packages/web/src/app/(app)/layout.tsx
@@ -33,6 +33,7 @@ import { ServiceErrorException } from "@/lib/serviceError";
 import { ConnectAccountsCard } from "@/ee/features/sso/components/connectAccountsCard";
 import { SidebarProvider } from "@/components/ui/sidebar";
 import { AppSidebar } from "./components/appSidebar";
+import { SidebarOverrideProvider } from "./components/appSidebar/sidebarOverrideContext";
 import { getUserChatHistory } from "@/features/chat/actions";
 
 interface LayoutProps {
@@ -184,16 +185,18 @@ export default async function Layout(props: LayoutProps) {
                 ) : null
             }
             <div className="fixed inset-0 flex bg-shell">
-                <SidebarProvider defaultOpen={cookieStore.get("sidebar_state")?.value !== "false"}>
-                    <AppSidebar session={session} chatHistory={chatHistory} />
-                    <div className="flex-1 min-h-0 flex flex-col pt-2 pb-2 pr-2">
-                        <div className="flex-1 min-h-0 bg-background flex flex-col border border-[#1d1d1f] rounded-xl">
-                            <div className="flex-1 min-h-0 overflow-y-auto">
-                                {children}
+                <SidebarOverrideProvider>
+                    <SidebarProvider defaultOpen={cookieStore.get("sidebar_state")?.value !== "false"}>
+                        <AppSidebar session={session} chatHistory={chatHistory} />
+                        <div className="flex-1 min-h-0 flex flex-col pt-2 pb-2 pr-2">
+                            <div className="flex-1 min-h-0 bg-background flex flex-col border border-[#1d1d1f] rounded-xl">
+                                <div className="flex-1 min-h-0 overflow-y-auto">
+                                    {children}
+                                </div>
                             </div>
                         </div>
-                    </div>
-                </SidebarProvider>
+                    </SidebarProvider>
+                </SidebarOverrideProvider>
             </div>
             <SyntaxReferenceGuide />
             <GitHubStarToast />
