+++ packages/web/src/app/(app)/settings/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Metadata } from "next"
-import { SidebarNav } from "./components/sidebar-nav"
+import { SettingsSidebarOverride, SidebarNavGroup, SidebarNavItem } from "./components/settingsSidebarOverride"
 import { redirect } from "next/navigation";
 import { auth } from "@/auth";
 import { isServiceError } from "@/lib/utils";
@@ -30,25 +30,21 @@ export default async function SettingsLayout(
         return redirect('/');
     }
 
-    const sidebarNavItems = await getSidebarNavItems();
-    if (isServiceError(sidebarNavItems)) {
-        throw new ServiceErrorException(sidebarNavItems);
+    const sidebarNavGroups = await getSidebarNavGroups();
+    if (isServiceError(sidebarNavGroups)) {
+        throw new ServiceErrorException(sidebarNavGroups);
     }
 
     return (
         <div>
+            <SettingsSidebarOverride groups={sidebarNavGroups} />
             <main className="flex justify-center p-4">
                 <div className="w-full max-w-6xl rounded-lg p-6">
                     <div className="container mx-auto">
                         <div className="mb-16">
                             <h1 className="text-3xl font-semibold">Settings</h1>
                         </div>
-                        <div className="flex flex-row gap-10">
-                            <aside className="lg:w-48">
-                                <SidebarNav items={sidebarNavItems} />
-                            </aside>
-                            <div className="w-full rounded-lg">{children}</div>
-                        </div>
+                        <div className="w-full rounded-lg">{children}</div>
                     </div>
                 </div>
             </main>
@@ -56,7 +52,7 @@ export default async function SettingsLayout(
     )
 }
 
-export const getSidebarNavItems = async () =>
+export const getSidebarNavGroups = async () =>
     withAuth(async ({ role }) => {
         let numJoinRequests: number | undefined;
         if (role === OrgRole.OWNER) {
@@ -72,49 +68,63 @@ export const getSidebarNavItems = async () =>
             throw new ServiceErrorException(connectionStats);
         }
 
-        return [
-            ...(role === OrgRole.OWNER ? [
-                {
-                    title: "Access",
-                    href: `/settings/access`,
-                }
-            ] : []),
-            ...(role === OrgRole.OWNER ? [{
-                title: "Members",
-                isNotificationDotVisible: numJoinRequests !== undefined && numJoinRequests > 0,
-                href: `/settings/members`,
-            }] : []),
-            ...(role === OrgRole.OWNER ? [
-                {
-                    title: "Connections",
-                    href: `/settings/connections`,
-                    hrefRegex: `/settings/connections(/[^/]+)?$`,
-                    isNotificationDotVisible: connectionStats.numberOfConnectionsWithFirstTimeSyncJobsInProgress > 0,
-                }
-            ] : []),
-            ...(env.DISABLE_API_KEY_USAGE_FOR_NON_OWNER_USERS === 'false' || role === OrgRole.OWNER ? [
-                {
-                    title: "API Keys",
-                    href: `/settings/apiKeys`,
-                }
-            ] : []),
-            ...(role === OrgRole.OWNER ? [
-                {
-                    title: "Analytics",
-                    href: `/settings/analytics`,
-                },
-            ] : []),
-            ...(hasEntitlement("sso") ? [
-                {
-                    title: "Linked Accounts",
-                    href: `/settings/linked-accounts`,
-                }
-            ] : []),
-            ...(role === OrgRole.OWNER ? [
-                {
-                    title: "License",
-                    href: `/settings/license`,
-                }
-            ] : []),
-        ]
+        const groups: SidebarNavGroup[] = [
+            {
+                label: "Account",
+                items: [
+                    ...(hasEntitlement("sso") ? [
+                        {
+                            title: "Linked Accounts",
+                            href: `/settings/linked-accounts`,
+                            icon: "link" as const,
+                        }
+                    ] : []),
+                    ...(env.DISABLE_API_KEY_USAGE_FOR_NON_OWNER_USERS === 'false' || role === OrgRole.OWNER ? [
+                        {
+                            title: "API Keys",
+                            href: `/settings/apiKeys`,
+                            icon: "key-round" as const,
+                        }
+                    ] : []),
+                ],
+            },
+        ];
+
+        if (role === OrgRole.OWNER) {
+            groups.push({
+                label: "Workspace",
+                items: [
+                    {
+                        title: "Access",
+                        href: `/settings/access`,
+                        icon: "shield" as const,
+                    },
+                    {
+                        title: "Members",
+                        isNotificationDotVisible: numJoinRequests !== undefined && numJoinRequests > 0,
+                        href: `/settings/members`,
+                        icon: "users" as const,
+                    },
+                    {
+                        title: "Connections",
+                        href: `/settings/connections`,
+                        hrefRegex: `/settings/connections(/[^/]+)?$`,
+                        isNotificationDotVisible: connectionStats.numberOfConnectionsWithFirstTimeSyncJobsInProgress > 0,
+                        icon: "plug" as const,
+                    },
+                    {
+                        title: "Analytics",
+                        href: `/settings/analytics`,
+                        icon: "chart-area" as const,
+                    },
+                    {
+                        title: "License",
+                        href: `/settings/license`,
+                        icon: "scroll-text" as const,
+                    },
+                ],
+            });
+        }
+
+        return groups.filter(g => g.items.length > 0);
     });
