+++ packages/web/src/components/ui/sidebar.tsx
@@ -106,7 +106,8 @@ const SidebarProvider = React.forwardRef<
     React.useEffect(() => {
       const handleKeyDown = (event: KeyboardEvent) => {
         if (
-          event.key === SIDEBAR_KEYBOARD_SHORTCUT
+          event.key === SIDEBAR_KEYBOARD_SHORTCUT &&
+          !(event.metaKey || event.ctrlKey)
         ) {
           event.preventDefault()
           toggleSidebar()
